modified announcement and updateUser controller

+++ controllers/announcement.controllers.mjs
@@ -6,31 +6,45 @@ const getAllAnnouncement = asyncHandler(async (request, response) => {
     if (request.user.role === 'student') {
         const level = request.user.level;
 
-        const announcements = await Announcement.find({ level }).populate('createdBy', 'name email')
+        const announcements = await Announcement.find({ level }).populate('createdBy', 'name')
         if (announcements.length === 0) {
             response.status(404)
             throw new Error(`Announcements are yet to made`)
         }
         response.status(200).json({
-            success: true,
-        count: announcements.length,
+            count: announcements.length,
             announcements
         })
     }
 
-    const announcements = await Announcement.find()
-    if (announcements.length === 0) {
-        response.status(404)
-        throw new Error(`Announcements are yet to made`)
+    if (request.user.role === 'teacher') {
+        const announcements = await Announcement.find({ createdBy: request.user._id })
+        if (announcements.length === 0) {
+            response.status(404)
+            throw new Error(`You have not created any announcement yet.`)
+        }
+        response.status(200).json({
+            count: announcements.length,
+            announcements
+        })
     }
-    response.status(200).json({
-        announcements
-    })
+
+    if (request.user.role === 'admin') {
+        const announcements = await Announcement.find().populate('createdBy', 'name email')
+        if (announcements.length === 0) {
+            response.status(404)
+            throw new Error(`Announcements are yet to made.`)
+        }
+        response.status(200).json({
+            count: announcements.length,
+            announcements
+        })
+    }
 })
 
 const createAnnouncement = asyncHandler(async (request, response) => {
     const { title, content, level } = request.body
-    if( !title || !content || !level ){
+    if (!title || !content || !level) {
         response.status(400)
         throw new Error(`Input all fields`)
     }
@@ -46,18 +60,23 @@ const updateAnnouncement = asyncHandler(async (request, response) => {
 
     const { id } = request.params
     const { title, content, level } = request.body
-    if( !title || !content || !level ){
+    if (!title || !content || !level) {
         response.status(400)
         throw new Error(`Input all fields`)
     }
 
     const announcement = await Announcement.findById(id)
-    if(!announcement){
+    if (!announcement) {
         response.status(404)
         throw new Error(`Announcement not found`)
     }
 
-    const updatedAnnouncement = await Announcement.findByIdAndUpdate( id, { title, content, level }, {new: true} )
+    if (announcement.createdBy.toString() !== request.user._id.toString()) {
+        response.status(403)
+        throw new Error(`Unauthorized, you are not the author of this announcement`)
+    }
+
+    const updatedAnnouncement = await Announcement.findByIdAndUpdate(id, { title, content, level }, { new: true })
     response.status(200).json({
         message: `Announcement updated successfully`,
         announcement: updatedAnnouncement
@@ -66,13 +85,18 @@ const updateAnnouncement = asyncHandler(async (request, response) => {
 
 const deleteAnnouncement = asyncHandler(async (request, response) => {
     const announcement = await Announcement.findById(request.params.id)
-    if(!announcement){
+    if (!announcement) {
         response.status(404)
         throw new Error(`Announcement not found`)
     }
 
+    if (announcement.createdBy.toString() !== request.user._id.toString()) {
+        response.status(403)
+        throw new Error(`Unauthorized, you are not the author of this announcement`)
+    }
+
     const announcementDeleted = await Announcement.findByIdAndDelete(request.params.id)
-    if(!announcementDeleted){
+    if (!announcementDeleted) {
         response.status(500)
         throw new Error(`Error deleting announcement`)
     }
